requirements.txt
streamlit
pacuare
python-dotenv